+++ src/test/java/ortus/boxlang/runtime/operators/PlusTest.java
@@ -36,14 +36,14 @@ public class PlusTest {
 	@Test
 	void testItCanAddNumbers() {
 		assertThat( Plus.invoke( 3, 2 ) ).isEqualTo( 5 );
-		assertThat( Plus.invoke( 3.5, 2.5 ) ).isEqualTo( 6 );
+		assertThat( Plus.invoke( 3.5, 2.5 ).doubleValue() ).isEqualTo( 6 );
 	}
 
 	@DisplayName( "It can add strings" )
 	@Test
 	void testItCanAddStrings() {
 		assertThat( Plus.invoke( "3", "2" ) ).isEqualTo( 5 );
-		assertThat( Plus.invoke( "3.5", "2.5" ) ).isEqualTo( 6 );
+		assertThat( Plus.invoke( "3.5", "2.5" ).doubleValue() ).isEqualTo( 6 );
 	}
 
 	@DisplayName( "It can compound add" )
